ai_service/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
joblib
prometheus-client
python-jose[cryptography]
sqlalchemy
pydantic
python-dotenv
